# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation de l'API du module de billeterie automatique",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/all_marchandises": {
            "get": {
                "tags": [
                    "The_marchandise"
                ],
                "summary": "Afficher la liste des marchandises",
                "description": "Renvoyer la liste de tous les marchandises",
                "operationId": "6c5d162cad59f50bad1261ed66e5bc20",
                "responses": {
                    "200": {
                        "description": "Les marchandises ont été trouvé et leurs informations sont retournées",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Aucune marchandise dans la base de données ",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/marchandise_show/{id}": {
            "get": {
                "tags": [
                    "The_marchandise"
                ],
                "summary": "Afficher les infos d'une marchandise",
                "description": "Renvoyer les infos d'une marchandise",
                "operationId": "e3912e798d1cbbd268565f49f63d1c64",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Identifiant d'une marchandise à voir ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " trouvé  !",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun  ne correspond à l'identifiant saisi",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/marchandise_store": {
            "post": {
                "tags": [
                    "The_marchandise"
                ],
                "summary": "Enregistrer une nouvelle marchandise",
                "operationId": "e8df4a61c4b3a39316255c571b152856",
                "parameters": [
                    {
                        "name": "info",
                        "in": "query",
                        "description": "Information sur la marchandise",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nbr_colis",
                        "in": "query",
                        "description": "Nombre de colis.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type_colis",
                        "in": "query",
                        "description": "Type dee colis.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "catégorie",
                        "in": "query",
                        "description": "Categorie d'apartenance du colis.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " enregistre avec succes.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Echec d'enregistrement de l'.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "402": {
                        "description": "Un client avec ces informations existe déja",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/marchandise_destroy/{id}": {
            "delete": {
                "tags": [
                    "The_marchandise"
                ],
                "summary": "Supprimer une marchandise",
                "description": "Elle se charger de supprimer un  et donc son compte",
                "operationId": "11dd82f834df889b8827c18ee6d72d70",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Identifiant du marchandise à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " supprimé avec succès!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "L'étudiant n'existe pas !",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    }
}